// src/types/pipeline.ts
export interface ParameterDefinition {
  name: string;
  type:
    | 'string'
    | 'number'
    | 'boolean'
    | 'object'
    | 'step'
    | 'stepList'
    | 'job'
    | 'jobList'
    | 'stage'
    | 'stageList';
  default?: unknown;
  displayName?: string;
  values?: unknown[];
}